feat(experiences-sdk-react)!: update basic component IDs with contentful prefix
Any existing experiences using the basic components will need to be updated. To
do so, import and run the `maintainBasicComponentIdsWithoutPrefix` method and
run it early on in your application (before the first `ExperienceRoot` is
rendered) like so:

```tsx
import { maintainBasicComponentIdsWithoutPrefix } from '@contentful/experiences-sdk-react';
maintainBasicComponentIdsWithoutPrefix();
```

This will register a set of the basic components with the older ids and allow
your experiences to continue to work. To migrate to the new components, replace
any of the components marked with "[OLD]" in the editor with their newer
counterparts (without the '[OLD]').

The `maintainBasicComponentIdsWithoutPrefix` is meant to help older experiences
built on a early alpha release to migrate and will be removed in the next major
release.